+++ go/test/endtoend/vtorc/readtopologyinstance/main_test.go
@@ -30,9 +30,9 @@ import (
 	"vitess.io/vitess/go/vt/vtorc/server"
 
 	_ "github.com/go-sql-driver/mysql"
-	_ "github.com/mattn/go-sqlite3"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	_ "modernc.org/sqlite"
 )
 
 func TestReadTopologyInstanceBufferable(t *testing.T) {
