+++ go/vt/external/golib/sqlutils/dialect.go
