+++ go/vt/vtgr/db/db.go
@@ -27,9 +27,9 @@ import (
 	"sync"
 	"time"
 
+	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/log"
 	"vitess.io/vitess/go/vt/vtgr/config"
-	"vitess.io/vitess/go/vt/vtgr/external/golib/sqlutils"
 )
 
 var (
