+++ go/vt/vtgr/db/mysql.go
@@ -26,11 +26,12 @@ import (
 	gouuid "github.com/google/uuid"
 	"github.com/spf13/pflag"
 
+	"vitess.io/vitess/go/vt/external/golib/sqlutils"
+
 	"vitess.io/vitess/go/mysql"
 	"vitess.io/vitess/go/vt/log"
 	"vitess.io/vitess/go/vt/servenv"
 	"vitess.io/vitess/go/vt/vtgr/config"
-	"vitess.io/vitess/go/vt/vtgr/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/vtgr/inst"
 )
