+++ go/vt/vtorc/attributes/attributes_dao.go
@@ -19,10 +19,8 @@ package attributes
 import (
 	"fmt"
 
+	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/log"
-
-	"github.com/openark/golib/sqlutils"
-
 	"vitess.io/vitess/go/vt/vtorc/db"
 )
