+++ go/vt/vtorc/db/db.go
@@ -20,8 +20,7 @@ import (
 	"database/sql"
 	"strings"
 
-	"github.com/openark/golib/sqlutils"
-
+	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/log"
 	"vitess.io/vitess/go/vt/vtorc/config"
 )
