+++ go/vt/vtorc/inst/analysis_dao_test.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/openark/golib/sqlutils"
+	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 
 	topodatapb "vitess.io/vitess/go/vt/proto/topodata"
 	"vitess.io/vitess/go/vt/vtorc/db"
