+++ go/vt/vtorc/inst/candidate_database_instance_dao.go
@@ -17,10 +17,8 @@
 package inst
 
 import (
-	"github.com/openark/golib/sqlutils"
-
+	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/log"
-
 	"vitess.io/vitess/go/vt/vtorc/config"
 	"vitess.io/vitess/go/vt/vtorc/db"
 )
