+++ go/vt/vtorc/inst/resolve_dao.go
@@ -19,10 +19,9 @@ package inst
 import (
 	"github.com/rcrowley/go-metrics"
 
+	"vitess.io/vitess/go/vt/external/golib/sqlutils"
+
 	"vitess.io/vitess/go/vt/log"
-
-	"github.com/openark/golib/sqlutils"
-
 	"vitess.io/vitess/go/vt/vtorc/config"
 	"vitess.io/vitess/go/vt/vtorc/db"
 )
