+++ go/vt/vtorc/inst/tablet_dao.go
@@ -20,13 +20,12 @@ import (
 	"context"
 	"errors"
 
+	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/log"
 
 	"google.golang.org/protobuf/encoding/prototext"
 
 	"google.golang.org/protobuf/proto"
-
-	"github.com/openark/golib/sqlutils"
 
 	"vitess.io/vitess/go/vt/logutil"
 	replicationdatapb "vitess.io/vitess/go/vt/proto/replicationdata"
