+++ go/vt/vtorc/process/health_dao.go
@@ -19,10 +19,8 @@ package process
 import (
 	"time"
 
+	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/log"
-
-	"github.com/openark/golib/sqlutils"
-
 	"vitess.io/vitess/go/vt/vtorc/config"
 	"vitess.io/vitess/go/vt/vtorc/db"
 )
